api/requirements.txt
# API Requirements (minimal for production)
tensorflow==2.15.0
numpy==1.24.3
pillow==10.2.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Optional: ONNX for optimized inference
onnxruntime==1.16.3